Removed legacy functions from src/memstack.c and include/memstack.h
After the changes of the last two commits, src/memstack.c and include/memstack.h have been modified significantly, and the previous code within these files were moved to the bottom and commented out.
This was so the original functions did not affect the new functions being implemented (some functions had the same name).
It is not out of disrespect, but out of convince to remove the commented out code. It looks messy and could confuse others looking at the code.

changelog:
- Removed original code commented at the bottom of src/memstack.c
- Removed original code commented at the bottom of header/memstack.h