Fixed dead memstack node allocated every msnew()
In msnew() after creating a new memsteck we also create the first memstack_chain_ptr* node.
However, this node is never used as msalloc() creates a new memstack_chain_ptr* node on it's own.
This is a quick fix, but also required a check in msalloc() to check if a first node exists or not.

changelog:
- Removed dead node from the beginning of every memstack