Added msprint(), mspop(), and mspush()
msprint(memstack*) is for debugging uses. It displays all the nodes in a memstack.
mspush(memstack*, void*) adds a new memstack node using memory provided by the user.
This is useful if the user wants to add new allocations to the memstack but does not have control over the allocations (like when using a library).
mspop(memstack*) will free the last node of the memstack and return the user allocated memory. This is useful if the user does no longer want an allocation to be freed by msfree() or msclear().

changelog:
- Added new function mspop()
- Added new function mspush()
- Added new function msprint()