Fixed bug with mspop() undefined behaviour
Added a length to memstack to help keep track of the number of nodes. Length is increased or decreased through functions that manipulate the memstack.
When memstack*->length is less than 2 mspop() produces undefined behaviour due to the dereferencing of NULL.
This commit fixes this issue by checking the memstack length and adjusting its de-allocation.

changelog:
- Added a length attribute to memstack struct
- Made the memstack length increase or decrease by the functions allocating or deallocating the memstack_chain_ptr* chain
- Fixed issue with mspop() undefined behaviour when given a memstack that doesn't hold more than 1 memstack_chain_ptr*