Updated description and removed usage info for now
I’ve change the description to be more accurate to what memstack is (I.e. it is a library not really a tool).
I’ve added more detailed instructions to the “Build from source” part of the installation section.
I’ve removed a large portion of the usage content: how to compile, what each function does, examples. 
I plan on rework the sections removed in the future.

changelog:
- Updated description
- Removed “usage” section
- Updated “Build from source” instructions